objects: implement wl_surface, wl_region, linux-dmabuf
wlr_surface will not be used in taiwins as it is not a independent component. So
now we have to implement most of our own wayland objects, especially wl_surface
(as the core component in wayland system).

The implementation is far from complete and there would be quite a few bugs. For
example, subsurface will now work at all. But right now I would need to focus on
getting standalone wl_surface to work.

Signed-off-by: Xichen Zhou <xzhou@xeechou.net>