seat: fix tw_seat_client with multiple resource xdg_shell: releative fix to xdg_shell_surface tw_seat: add version check to wl_resources
For multi-process application like firefox. They often bind seats/xdg_wm_base
multiple times(with the same client). For wm_base this may not be a huge problem
since we are not navitgating wm_base based on client.

The `tw_seat_client` here however has trouble navigating resources (maybe we
will also have in data-device) because we find `tw_seat_client` by a
`wl_client`. It would not give us the right `tw_seat_client`. The result is that
we cannot send out input messages to the right `wl_seat.resource`. Because of
this, free `tw_seat_client` becomes tricky, since we free it by destroying all the
resource. When we destroy the last seat_client, the head will be gone.

having a big tw_xdg_surface is probably not a great idea. The pointer chasing is
more and more a obvious problem now.

many clients does not implement the some of the listeners. Checking version is
vital apparently.

Signed-off-by: Xichen Zhou <xzhou@xeechou.net>