desktop: resolve the jittering moving/resizing problems in floating layout
moving problem was due to incorrect coordinates mapping. In fact, once we start
the grab, the sx, sy should never change. Using gx, gy for getting deltas works
as well.

The resizing problem has one more depth to it. Altough we changed geometry of
the surface and send to clients, but it may not be received before next
resizing, for that, we would need to work on the "planed size" instead of using
the current surface size.

Signed-off-by: Xichen Zhou <xzhou@xeechou.net>