render: handle render_output destroy on exit
When we exiting compositor, render_output would unset context, which in turn
cause every wl_surfaces to re-assign outputs. Previously wl_surfaces would
assume there is always an output to assign but it is not the case.

This fix only handle this issue, we need to truly tackle the output lost
problem for wl_surfaces.

Signed-off-by: Xichen Zhou <xzhou@xeechou.net>